git: Add function to toggle fullscreen magit status
This patch resolves issue #7470.

Currently, if git-magit-status-fullscreen is true and the
magit-display-buffer-function is changed, there's no way to easily change it
back if you don't want magit status in a fullscreen buffer.

This patch adds a function spacemacs/magit-status-toggle-fullscreen which allows
one to easily toggle a fullscreen magit status buffer.